pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "membench"
version = "0.1.0"
description = "Memory Benchmark Suite for Long-Term Agent Memory Evaluation"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Memory Benchmark Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
# Core memory libraries
mem0 = ["mem0ai>=1.0.11"]
zep = ["zep-python>=2.0.2"]
graphiti = ["graphiti-core>=0.28.2", "neo4j>=5.26.0"]
letta = ["letta>=0.16.7"]
mempalace = ["mempalace>=3.1.0", "chromadb>=0.5.0,<0.7"]
vector = ["chromadb>=0.4.0", "faiss-cpu>=1.8.0"]

# All memory stores
all = [
    "mem0ai>=1.0.11",
    "zep-python>=2.0.2",
    "graphiti-core>=0.28.2",
    "neo4j>=5.26.0",
    "letta>=0.16.7",
    "mempalace>=3.1.0",
    "chromadb>=0.5.0,<0.7",
    "faiss-cpu>=1.8.0",
]

# Development
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "sphinx>=7.0.0",
    "myst-parser>=2.0.0",
]

[project.scripts]
membench = "membench.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/membench"
Documentation = "https://github.com/yourusername/membench/blob/main/docs/DOCUMENTATION.md"
Repository = "https://github.com/yourusername/membench"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
